requirements.txt
google-generativeai>=0.8.6
google-auth>=2.29.0
google-api-python-client>=2.120.0
google-cloud-secret-manager>=2.20.0
gspread>=6.0.0
requests>=2.31.0
pdfplumber>=0.11.0
python-docx>=1.1.0
typing_extensions>=4.0.0